merge new commit from alibaba to my dataagent

+++ data-agent-management/src/main/java/com/alibaba/cloud/ai/dataagent/entity/AgentDatasource.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2024-2025 the original author or authors.
+ * Copyright 2024-2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * https://www.apache.org/licenses/LICENSE-2.0
+ *     https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,18 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.alibaba.cloud.ai.dataagent.entity;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import org.springframework.format.annotation.DateTimeFormat;
-
 import java.time.LocalDateTime;
 import java.util.List;
-
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
+import org.springframework.format.annotation.DateTimeFormat;
 
 @Data
 @NoArgsConstructor
@@ -33,7 +30,7 @@ public class AgentDatasource {
 
 	private Integer id;
 
-	private Integer agentId;
+	private Long agentId;
 
 	private Integer datasourceId;
 
@@ -53,7 +50,7 @@ public class AgentDatasource {
 	// 当前数据源选中的表
 	private List<String> selectTables;
 
-	public AgentDatasource(Integer agentId, Integer datasourceId) {
+	public AgentDatasource(Long agentId, Integer datasourceId) {
 		this.agentId = agentId;
 		this.datasourceId = datasourceId;
 		this.isActive = 1;
